requirements.txt
attrs==18.1.0
certifi==2018.4.16
docutils==0.14
enum34==1.1.6
future==0.16.0
lockfile==0.12.2
more-itertools==4.1.0
netifaces==0.10.7
pluggy==0.6.0
psutil==5.4.5
py==1.5.3
PySocks==1.6.8
pytest==3.5.1
python-daemon==2.1.2
python-telegram-bot==10.1.0
pyvalid==0.9
six==1.11.0
